@@ -2,29 +2,24 @@
   "type": "module",
   "devDependencies": {
     "@jest/globals": "29.7.0",
-    "@types/chrome": "0.0.278",
-    "@types/jest": "29.5.13",
-    "@types/node": "22.7.5",
-    "css-loader": "7.1.2",
-    "esbuild": "0.24.0",
-    "esbuild-loader": "4.2.2",
+    "@types/chrome": "0.0.299",
+    "@types/jest": "29.5.14",
+    "@types/node": "22.10.9",
+    "esbuild": "0.24.2",
+    "esbuild-svelte": "0.9.0",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
-    "mini-css-extract-plugin": "2.9.1",
-    "prettier": "3.3.3",
-    "sass": "1.77.8",
-    "svelte": "4.2.19",
-    "svelte-check": "4.0.2",
-    "svelte-loader": "3.2.4",
+    "prettier": "3.4.2",
+    "sass": "1.83.4",
+    "sv": "0.6.15",
+    "svelte": "5.19.2",
+    "svelte-check": "4.1.4",
     "svelte-preprocess": "6.0.3",
     "ts-jest": "29.2.5",
     "ts-node": "10.9.2",
-    "tsx": "4.19.1",
-    "typescript": "5.6.3",
-    "webpack": "5.95.0",
-    "webpack-cli": "5.1.4"
+    "typescript": "5.7.3"
   },
   "dependencies": {
-    "@noble/hashes": "1.5.0"
+    "@noble/hashes": "1.7.1"
   }
 }
