@@ -1,30 +1,34 @@
 <script lang="ts">
-  import { createEventDispatcher } from 'svelte';
+  import type { Snippet } from 'svelte';
 
-  export let title: string = '';
-  export let description: string = '';
+  let {
+    title = '',
+    description = '',
+    eventClose: closeEvent,
+    children,
+  }: {
+    title: string;
+    description: string;
+    eventClose?: () => void;
+    children?: Snippet;
+  } = $props();
+  let showContent: boolean = $state(false);
   let selfEl: HTMLDialogElement | null = null;
-  let showContent: boolean = false;
-  const dispatch = createEventDispatcher();
 
-  export function showModal() {
+  export function show() {
     selfEl?.showModal();
     document.addEventListener('keydown', onKeyboardEvent, { capture: true });
-    selfEl?.addEventListener('click', onWindowClick);
+    selfEl?.addEventListener('click', onSelfClick);
     showContent = true;
   }
 
-  function onClose() {
-    document.removeEventListener('keydown', onKeyboardEvent, { capture: true });
-    selfEl?.removeEventListener('click', onWindowClick);
-    showContent = false;
+  export function hide() {
     selfEl?.close();
-    dispatch('closeDialog');
   }
 
-  function onWindowClick(e: MouseEvent) {
+  function onSelfClick(e: MouseEvent) {
     if (e.currentTarget === e.target) {
-      onClose();
+      hide();
     }
   }
 
@@ -32,34 +36,41 @@
     if (e.key === 'Escape') {
       e.preventDefault();
       e.stopImmediatePropagation();
-      onClose();
+      hide();
     }
+  }
+
+  function onClose() {
+    document.removeEventListener('keydown', onKeyboardEvent, { capture: true });
+    selfEl?.removeEventListener('click', onSelfClick);
+    showContent = false;
+    closeEvent?.();
   }
 </script>
 
-<dialog bind:this={selfEl}>
+<dialog bind:this={selfEl} onclose={onClose}>
+  <header>
+    <div class="title">
+      {title}
+      {#if description}
+        <div class="description">{description}</div>
+      {/if}
+    </div>
+
+    <a
+      title="Close"
+      aria-label="Close"
+      class="close-icon"
+      href="void(0)"
+      onclick={(e) => {
+        e.preventDefault();
+        hide();
+      }}><span class="icon -remove"></span></a
+    >
+  </header>
+
   {#if showContent}
-    <header>
-      <div class="title">
-        {#if title}
-          {title}
-        {/if}
-        {#if description}
-          <div class="description">{description}</div>
-        {/if}
-      </div>
-
-      <a
-        title="Close"
-        class="close-icon"
-        href="void(0)"
-        on:click|preventDefault={onClose}
-      >
-        <span class="icon -remove" />
-      </a>
-    </header>
-
-    <slot />
+    {@render children?.()}
   {/if}
 </dialog>
 
