@@ -1,7 +1,7 @@
 <script lang="ts">
-  import { portPost, type TMsgMediaCommand } from '@/api/communication.ts';
+  import { portPost, type TMsgMediaCommand } from '../../api/communication.ts';
 
-  export let mediaId: string;
+  let { mediaId }: { mediaId: string } = $props();
 
   function onMediaCommand(cmd: TMsgMediaCommand['cmd']) {
     portPost({ msg: 'media-command', mediaId, cmd });
@@ -10,44 +10,51 @@
 
 <span class="media-commands">
   <button
-    on:click={() => void onMediaCommand('log')}
+    onclick={() => void onMediaCommand('log')}
     title="console.log(media)"
+    aria-label="console.log(media)"><span class="icon -console"></span></button
   >
-    <span class="icon -console" />
-  </button>
   <button
-    on:click={() => void onMediaCommand('locate')}
+    onclick={() => void onMediaCommand('locate')}
     title="media.scrollIntoView()"
+    aria-label="media.scrollIntoView()"
+    ><span class="icon -locate"></span></button
   >
-    <span class="icon -locate" />
-  </button>
-  <button on:click={() => void onMediaCommand('load')} title="media.load()">
-    <span class="icon -refresh" />
-  </button>
-  <button on:click={() => void onMediaCommand('play')} title="media.play()">
-    <span class="icon -play" />
-  </button>
-  <button on:click={() => void onMediaCommand('pause')} title="media.pause()">
-    <span class="icon -pause" />
-  </button>
   <button
-    on:click={() => void onMediaCommand('frame-backward')}
+    onclick={() => void onMediaCommand('load')}
+    title="media.load()"
+    aria-label="media.load()"><span class="icon -refresh"></span></button
+  >
+  <button
+    onclick={() => void onMediaCommand('play')}
+    title="media.play()"
+    aria-label="media.play()"><span class="icon -play"></span></button
+  >
+  <button
+    onclick={() => void onMediaCommand('pause')}
+    title="media.pause()"
+    aria-label="media.pause()"><span class="icon -pause"></span></button
+  >
+  <button
+    onclick={() => void onMediaCommand('frame-backward')}
     title="Seek -16ms"
+    aria-label="Seek -16ms"><span class="icon -frame-backward"></span></button
   >
-    <span class="icon -frame-backward" />
-  </button>
   <button
-    on:click={() => void onMediaCommand('frame-forward')}
+    onclick={() => void onMediaCommand('frame-forward')}
     title="Seek +16ms"
+    aria-label="Seek +16ms"><span class="icon -frame-forward"></span></button
   >
-    <span class="icon -frame-forward" />
-  </button>
-  <button on:click={() => void onMediaCommand('slower')} title="Speed -10%">
-    <span class="icon -slower" />
-  </button>
-  <button on:click={() => void onMediaCommand('faster')} title="Speed +10%">
-    <span class="icon -faster" />
-  </button>
+  <button
+    onclick={() => void onMediaCommand('slower')}
+    title="Speed -10%"
+    aria-label="Speed -10%"><span class="icon -slower"></span></button
+  >
+  <button
+    onclick={() => void onMediaCommand('faster')}
+    title="Speed +10%"
+    aria-label="Speed +10%"><span class="icon -faster"></span></button
+  >
 </span>
 
 <style lang="scss">
