@@ -1,21 +1,17 @@
 <script lang="ts">
-  import Variable from '@/view/components/Variable.svelte';
-  import type { TMediaMetrics } from '@/api/mediaMonitor.ts';
-  import MediaCommands from '@/view/components/MediaCommands.svelte';
-  import { portPost } from '@/api/communication.ts';
-  import { MEDIA_ELEMENT_TOGGABLE_PROPS } from '@/api/const.ts';
+  import type { TMediaMetrics } from '../../api/mediaMonitor.ts';
+  import { portPost } from '../../api/communication.ts';
+  import { MEDIA_ELEMENT_TOGGABLE_PROPS } from '../../api/const.ts';
+  import Variable from './Variable.svelte';
+  import MediaCommands from './MediaCommands.svelte';
 
-  export let metrics: TMediaMetrics;
+  let { metrics }: { metrics: TMediaMetrics } = $props();
 
   function isToggable(property: string) {
     return MEDIA_ELEMENT_TOGGABLE_PROPS.has(property);
   }
 
   function onToggleBoolean(property: string) {
-    if (!isToggable(property)) {
-      return;
-    }
-
     portPost({
       msg: 'media-command',
       mediaId: metrics.mediaId,
@@ -23,60 +19,68 @@
       property: property as keyof HTMLMediaElement,
     });
   }
+
+  function propValueFilter(value: unknown) {
+    if (value && typeof value === 'object') {
+      return JSON.stringify(value);
+    }
+
+    return value;
+  }
 </script>
 
 <table class="group">
   <caption class="bc-invert ta-l">
-    <MediaCommands bind:mediaId={metrics.mediaId} />
+    <MediaCommands mediaId={metrics.mediaId} />
   </caption>
-  <tr>
-    <td class="events">
-      <table class="w-full">
-        <caption class="bc-invert ta-l">Events</caption>
-        {#each Object.entries(metrics.events) as [label, value] (label)}
-          <tr class:isPassive={0 === value} class:isActive={0 !== value}>
-            <td class="item-label">{label}</td>
-            <td class="item-value"><Variable bind:value /></td>
-          </tr>
-        {/each}
-      </table>
-    </td>
-    <td class="props">
-      <table class="w-full">
-        <caption class="bc-invert ta-l">Properties</caption>
-        {#each Object.entries(metrics.props) as [label, value] (label)}
-          <tr
-            class:isPassive={null === value ||
-              false === value ||
-              '' === value ||
-              0 === value}
-            class:isActive={true === value}
-          >
-            <td class="item-label">{label}</td>
-            <td class="item-value">
-              {#if isToggable(label)}
-                <i
-                  class="isToggable"
-                  role="button"
-                  tabindex="0"
-                  on:keydown={(e) => {
-                    if (e.key !== 'Enter' && e.key !== ' ') return;
-                    e.preventDefault();
-                    onToggleBoolean(label);
-                  }}
-                  on:click={() => void onToggleBoolean(label)}>{value}</i
-                >
-              {:else if ['networkState', 'readyState'].includes(label)}
-                <Variable bind:value />
-              {:else}
-                {value}
-              {/if}
-            </td>
-          </tr>
-        {/each}
-      </table>
-    </td>
-  </tr>
+  <tbody>
+    <tr>
+      <td class="events">
+        <table class="w-full">
+          <caption class="bc-invert ta-l">Events</caption>
+          <tbody>
+            {#each Object.entries(metrics.events) as [label, value] (label)}
+              <tr class:isPassive={0 === value} class:isActive={0 !== value}>
+                <td class="item-label">{label}</td>
+                <td class="item-value"><Variable {value} /></td>
+              </tr>
+            {/each}
+          </tbody>
+        </table>
+      </td>
+      <td class="props">
+        <table class="w-full">
+          <caption class="bc-invert ta-l">Properties</caption>
+          <tbody>
+            {#each Object.entries(metrics.props) as [label, value] (label)}
+              <tr class:isPassive={!value} class:isActive={true === value}>
+                <td class="item-label">{label}</td>
+                <td class="item-value">
+                  {#if isToggable(label)}
+                    <i
+                      class="isToggable"
+                      role="button"
+                      tabindex="0"
+                      onkeydown={(e) => {
+                        if (e.key !== 'Enter' && e.key !== ' ') return;
+                        e.preventDefault();
+                        onToggleBoolean(label);
+                      }}
+                      onclick={() => void onToggleBoolean(label)}>{value}</i
+                    >
+                  {:else if ['networkState', 'readyState'].includes(label)}
+                    <Variable {value} />
+                  {:else}
+                    {propValueFilter(value)}
+                  {/if}
+                </td>
+              </tr>
+            {/each}
+          </tbody>
+        </table>
+      </td>
+    </tr>
+  </tbody>
 </table>
 
 <style lang="scss">
