@@ -1,20 +1,23 @@
 <script lang="ts">
-  import type { TClearTimerHistory } from '@/api/wrappers.ts';
-  import Trace from '@/view/components/Trace.svelte';
-  import TraceDomain from '@/view/components/TraceDomain.svelte';
-  import Variable from '@/view/components/Variable.svelte';
+  import type { TClearTimerHistory } from '../../wrapper/TimerWrapper.ts';
+  import { msToHms } from '../../api/time.ts';
+  import Trace from './Trace.svelte';
+  import TraceDomain from './TraceDomain.svelte';
+  import Variable from './Variable.svelte';
+  import TraceBreakpoint from './TraceBreakpoint.svelte';
+  import TraceBypass from './TraceBypass.svelte';
 
-  export let metric: TClearTimerHistory;
+  let { metric }: { metric: TClearTimerHistory } = $props();
 </script>
 
-<tr class="t-zebra" class:bc-error={metric.hasError}>
-  <td><TraceDomain bind:traceDomain={metric.traceDomain} /></td>
+<tr class="t-zebra">
   <td class="wb-all">
-    <Trace bind:trace={metric.trace} bind:traceId={metric.traceId} />
+    <TraceDomain traceDomain={metric.traceDomain} />
+    <Trace trace={metric.trace} />
   </td>
-  <td class="ta-c">
-    <Variable bind:value={metric.calls} />
-  </td>
-  <td class="ta-c">{metric.handler}</td>
-  <td class="ta-r">{metric.delay}</td>
+  <td class="ta-c"><Variable value={metric.calls} /></td>
+  <td class="ta-c"><Variable value={metric.handler} /></td>
+  <td class="ta-r" title={msToHms(metric.delay)}>{metric.delay}</td>
+  <td><TraceBypass traceId={metric.traceId} /></td>
+  <td><TraceBreakpoint traceId={metric.traceId} /></td>
 </tr>
