@@ -1,159 +1,112 @@
 <script lang="ts">
-  import type { TClearTimerHistory, TSetTimerHistory } from '@/api/wrappers.ts';
-  import Variable from '@/view/components/Variable.svelte';
-  import Trace from '@/view/components/Trace.svelte';
-  import TraceDomain from '@/view/components/TraceDomain.svelte';
+  import type {
+    TClearTimerHistory,
+    TSetTimerHistory,
+  } from '../../wrapper/TimerWrapper.ts';
   import {
-    DEFAULT_SORT,
+    DEFAULT_SORT_SET_TIMERS,
     getSettings,
-    EHistorySortField,
     setSettings,
-    type EHistorySortFieldKeys,
-    type ESortOrderKeys,
-  } from '@/api/settings.ts';
-  import TimersHistoryCellSort from '@/view/components/TimersHistoryCellSort.svelte';
-  import TimersClearHistory from '@/view/components/TimersClearHistory.svelte';
-  import { compareByFieldOrder } from '@/api/comparator.ts';
-  import Dialog from '@/view/components/Dialog.svelte';
+    ESortOrder,
+  } from '../../api/settings.ts';
+  import { compareByFieldOrder } from '../../api/comparator.ts';
+  import Variable from './Variable.svelte';
+  import SortableColumn from './SortableColumn.svelte';
+  import TimersSetHistoryMetric from './TimersSetHistoryMetric.svelte';
 
-  export let caption: string;
-  export let metrics: TSetTimerHistory[];
-  export let clearTimeoutHistory: TClearTimerHistory[] | null;
-  export let clearIntervalHistory: TClearTimerHistory[] | null;
-
-  let field: EHistorySortFieldKeys = DEFAULT_SORT.timersHistoryField;
-  let order: ESortOrderKeys = DEFAULT_SORT.timersHistoryOrder;
-  let dialog: Dialog | null = null;
-
-  $: sortedMetrics = metrics.sort(compareByFieldOrder(field, order));
+  let {
+    metrics,
+    clearTimeoutHistory,
+    clearIntervalHistory,
+    caption,
+  }: {
+    metrics: TSetTimerHistory[];
+    clearTimeoutHistory: TClearTimerHistory[] | null;
+    clearIntervalHistory: TClearTimerHistory[] | null;
+    caption?: string;
+  } = $props();
+  let sortField = $state(DEFAULT_SORT_SET_TIMERS.field);
+  let sortOrder = $state(DEFAULT_SORT_SET_TIMERS.order);
+  let sortedMetrics = $derived.by(() =>
+    metrics.sort(compareByFieldOrder(sortField, sortOrder))
+  );
 
   getSettings().then((settings) => {
-    field = settings.sort.timersHistoryField;
-    order = settings.sort.timersHistoryOrder;
+    sortField = settings.sortSetTimers.field;
+    sortOrder = settings.sortSetTimers.order;
   });
 
-  function onChangeSort(
-    e: CustomEvent<{ field: EHistorySortFieldKeys; order: ESortOrderKeys }>
-  ) {
-    field = e.detail.field;
-    order = e.detail.order;
+  function onChangeSort(_field: string, _order: ESortOrder) {
+    sortField = <keyof TSetTimerHistory>_field;
+    sortOrder = _order;
+
     setSettings({
-      sort: {
-        timersHistoryField: field,
-        timersHistoryOrder: order,
+      sortSetTimers: {
+        field: $state.snapshot(sortField),
+        order: $state.snapshot(sortOrder),
       },
     });
   }
-
-  let clearTimerHistoryMetrics: TClearTimerHistory[] = [];
-
-  function onFindRegressors(regressors: string[] | null) {
-    if (!dialog || !regressors?.length) {
-      return;
-    }
-
-    for (let n = regressors.length - 1; n >= 0; n--) {
-      const traceId = regressors[n];
-      let record = clearTimeoutHistory?.find((r) => r.traceId === traceId);
-      record ??= clearIntervalHistory?.find((r) => r.traceId === traceId);
-      if (record) {
-        clearTimerHistoryMetrics.push(record);
-      }
-    }
-
-    if (clearTimerHistoryMetrics.length) {
-      dialog.showModal();
-    }
-  }
-
-  function onCloseDialog() {
-    clearTimerHistoryMetrics.splice(0);
-  }
 </script>
-
-<Dialog
-  bind:this={dialog}
-  on:closeDialog={onCloseDialog}
-  title="Places from which timer with current callstack was prematurely canceled"
-  description="The information is actual only on time of demand. For full coverage - requires both clearTimeout and clearInterval panels enabled."
->
-  <TimersClearHistory
-    caption="Canceled by"
-    bind:metrics={clearTimerHistoryMetrics}
-  />
-</Dialog>
 
 <table data-navigation-tag={caption}>
   <caption class="bc-invert ta-l">
     {caption}
-    <Variable bind:value={metrics.length} />
+    <Variable value={metrics.length} />
   </caption>
-  <tr>
-    <th class="shaft"></th>
-    <th class="w-full">Callstack</th>
-    <th class="ta-c">
-      <TimersHistoryCellSort
-        field={EHistorySortField.calls}
-        currentField={field}
-        currentFieldOrder={order}
-        on:changeSort={onChangeSort}>Called</TimersHistoryCellSort
-      >
-    </th>
-    <th class="ta-c">
-      <TimersHistoryCellSort
-        field={EHistorySortField.handler}
-        currentField={field}
-        currentFieldOrder={order}
-        on:changeSort={onChangeSort}>Handler</TimersHistoryCellSort
-      >
-    </th>
-    <th class="ta-r">
-      <TimersHistoryCellSort
-        field={EHistorySortField.delay}
-        currentField={field}
-        currentFieldOrder={order}
-        on:changeSort={onChangeSort}>Delay</TimersHistoryCellSort
-      >
-    </th>
-    <th class="shaft"></th>
-  </tr>
-
-  {#each sortedMetrics as metric (metric.traceId)}
-    <tr class="t-zebra" class:bc-error={metric.hasError}>
-      <td><TraceDomain bind:traceDomain={metric.traceDomain} /></td>
-      <td class="wb-all">
-        <Trace bind:trace={metric.trace} bind:traceId={metric.traceId} />
-      </td>
-      <td class="ta-c">
-        <Variable bind:value={metric.calls} />
-      </td>
-      <td class="ta-c">{metric.handler}</td>
-      <td class="ta-r">{metric.delay}</td>
-      <td>
-        {#if metric.canceledByTraceIds?.length}
-          <a
-            role="button"
-            title={`${metric.isOnline ? 'Scheduled. ' : ''}Canceled by ${metric.canceledByTraceIds?.length}`}
-            href="void(0)"
-            on:click|preventDefault={() =>
-              void onFindRegressors(metric.canceledByTraceIds)}
-          >
-            {#if metric.isOnline}
-              <span class="icon -scheduled -small" />
-            {:else}
-              <span class="icon -remove -small" />
-            {/if}
-          </a>
-        {:else if metric.isOnline}
-          <span title="Scheduled" class="icon -scheduled -small" />
-        {/if}
-      </td>
+  <tbody>
+    <tr>
+      <th class="w-full">Callstack</th>
+      <th class="ta-r">
+        <SortableColumn
+          field="selfTime"
+          currentField={sortField}
+          currentFieldOrder={sortOrder}
+          eventChangeSorting={onChangeSort}>Self</SortableColumn
+        >
+      </th>
+      <th class="ta-c">
+        <SortableColumn
+          field="calls"
+          currentField={sortField}
+          currentFieldOrder={sortOrder}
+          eventChangeSorting={onChangeSort}>Called</SortableColumn
+        >
+      </th>
+      <th class="ta-c">
+        <SortableColumn
+          field="handler"
+          currentField={sortField}
+          currentFieldOrder={sortOrder}
+          eventChangeSorting={onChangeSort}>Handler</SortableColumn
+        >
+      </th>
+      <th class="ta-r">
+        <SortableColumn
+          field="delay"
+          currentField={sortField}
+          currentFieldOrder={sortOrder}
+          eventChangeSorting={onChangeSort}>Delay</SortableColumn
+        >
+      </th>
+      <th>
+        <SortableColumn
+          field="online"
+          currentField={sortField}
+          currentFieldOrder={sortOrder}
+          eventChangeSorting={onChangeSort}>Set</SortableColumn
+        >
+      </th>
+      <th title="Bypass"><span class="icon -bypass"></span></th>
+      <th title="Breakpoint"><span class="icon -breakpoint"></span></th>
     </tr>
-  {/each}
-</table>
 
-<style lang="scss">
-  .shaft {
-    min-width: var(--small-icon-size);
-  }
-</style>
+    {#each sortedMetrics as metric (metric.traceId)}
+      <TimersSetHistoryMetric
+        {metric}
+        {clearTimeoutHistory}
+        {clearIntervalHistory}
+      />
+    {/each}
+  </tbody>
+</table>
