@@ -1,15 +1,15 @@
 <script lang="ts">
-  import { ETraceDomain, type ETraceDomainKeys } from '@/api/wrappers.ts';
+  import { ETraceDomain } from '../../wrapper/TraceUtil.ts';
 
-  export let traceDomain: ETraceDomainKeys;
+  let { traceDomain }: { traceDomain: ETraceDomain } = $props();
 </script>
 
 {#if traceDomain === ETraceDomain.SAME}
-  <span class="icon -small -trace-local" title="Same domain" />
+  <span class="icon -small -trace-local" title="Same domain"></span>
 {:else if traceDomain === ETraceDomain.EXTERNAL}
-  <span class="icon -small -trace-external" title="External domain" />
+  <span class="icon -small -trace-external" title="External domain"></span>
 {:else if traceDomain === ETraceDomain.EXTENSION}
-  <span class="icon -small -trace-extension" title="Local extension" />
+  <span class="icon -small -trace-extension" title="Local extension"></span>
 {:else if traceDomain === ETraceDomain.UNKNOWN}
   <span title="Unknown domain">❓</span>
 {/if}
