@@ -3,8 +3,8 @@ import {
   windowPost,
   windowListen,
   portListen,
-} from '@/api/communication.ts';
-import { getSettings, onSettingsChange } from '@/api/settings.ts';
+} from './api/communication.ts';
+import { getSettings, onSettingsChange } from './api/settings.ts';
 
 getSettings().then((settings) => {
   windowPost({ msg: 'settings', settings: settings });
