@@ -1,21 +1,19 @@
-import { windowListen, windowPost } from '@/api/communication.ts';
-import { IS_DEV } from '@/api/env.ts';
+import { windowListen, windowPost } from './api/communication.ts';
+import { IS_DEV } from './api/env.ts';
 import {
   TELEMETRY_FREQUENCY_1PS,
   TELEMETRY_FREQUENCY_30PS,
-} from '@/api/const.ts';
-import { callingOnce, Timer } from '@/api/time.ts';
+} from './api/const.ts';
+import { Timer } from './api/time.ts';
 import {
   collectMediaMetrics,
   meetMedia,
   doMediaCommand,
   type TMediaTelemetry,
-} from '@/api/mediaMonitor.ts';
-import { Wrapper, ETimerType, type TWrapperMetrics } from '@/api/wrappers.ts';
-import {
-  DEFAULT_SETTINGS,
-  panelsArrayToVisibilityMap,
-} from '@/api/settings.ts';
+} from './api/mediaMonitor.ts';
+import { Wrapper, type TWrapperMetrics } from './wrapper/Wrapper.ts';
+import { panelsArray2Map, type TPanelMap } from './api/settings.ts';
+import { ETimerType } from './wrapper/TimerWrapper.ts';
 
 export interface TMetrics {
   mediaMetrics: TMediaTelemetry;
@@ -23,88 +21,44 @@ export interface TMetrics {
   collectingStartTime: number;
 }
 
-const wrapper = new Wrapper();
-const wrapEvalOnce = callingOnce(() => wrapper.wrapEval());
-const wrapSetTimeoutOnce = callingOnce(() => wrapper.wrapSetTimeout());
-const wrapClearTimeoutOnce = callingOnce(() => wrapper.wrapClearTimeout());
-const wrapSetIntervalOnce = callingOnce(() => wrapper.wrapSetInterval());
-const wrapClearIntervalOnce = callingOnce(() => wrapper.wrapClearInterval());
-const wrapRequestAnimationFrameOnce = callingOnce(() =>
-  wrapper.wrapRequestAnimationFrame()
-);
-const wrapCancelAnimationFrameOnce = callingOnce(() =>
-  wrapper.wrapCancelAnimationFrame()
-);
-const wrapRequestIdleCallbackOnce = callingOnce(() =>
-  wrapper.wrapRequestIdleCallback()
-);
-const wrapCancelIdleCallbackOnce = callingOnce(() =>
-  wrapper.wrapCancelIdleCallback()
-);
-let panels = panelsArrayToVisibilityMap(DEFAULT_SETTINGS.panels);
-const eachSecond = new Timer(
-  () => {
-    meetMedia(document.querySelectorAll('video,audio'));
-  },
-  1e3,
-  { interval: true }
-);
+let panels: TPanelMap;
+let wrapper: Wrapper;
+const eachSecond = new Timer({ delay: 1e3, repetitive: true }, () => {
+  meetMedia(document.querySelectorAll('video,audio'));
+  wrapper.eachSecond();
+});
 const tick = new Timer(
+  { delay: TELEMETRY_FREQUENCY_1PS, repetitive: true },
   function apiMonitorPostMetric() {
     const now = Date.now();
 
-    const metrics: TMetrics = {
-      mediaMetrics: collectMediaMetrics(panels.media.visible),
-      wrapperMetrics: wrapper.collectWrapperMetrics(panels),
-      collectingStartTime: now,
-    };
-
-    windowPost({ msg: 'telemetry', metrics });
-  },
-  TELEMETRY_FREQUENCY_1PS,
-  { interval: true, animation: false, measurable: true }
+    windowPost({
+      msg: 'telemetry',
+      metrics: {
+        mediaMetrics: collectMediaMetrics(panels.media.visible),
+        wrapperMetrics: wrapper.collectMetrics(),
+        collectingStartTime: now,
+      },
+    });
+  }
 );
 
-function startObserve() {
-  stopObserve();
-  tick.start();
-  eachSecond.start();
-}
-
-function stopObserve() {
-  tick.stop();
-  eachSecond.stop();
-}
-
 windowListen((o) => {
-  if (o.msg === 'telemetry-acknowledged') {
-    // adaptive update-frequency
-    const ackTrafficDuration = Date.now() - o.timeSent;
-    const newDelay = (o.trafficDuration + ackTrafficDuration) * 3;
-    tick.delay = Math.max(TELEMETRY_FREQUENCY_30PS, newDelay);
+  if (o.msg === 'settings' && o.settings && typeof o.settings === 'object') {
+    if (!wrapper) {
+      wrapper = new Wrapper(panels);
+    }
+    panels = panelsArray2Map(o.settings.panels);
+    wrapper.setup(panels, o.settings);
   } else if (o.msg === 'start-observe') {
-    startObserve();
+    tick.start();
+    eachSecond.start();
   } else if (o.msg === 'stop-observe') {
-    stopObserve();
-  } else if (
-    o.msg === 'settings' &&
-    o.settings &&
-    typeof o.settings === 'object'
-  ) {
-    panels = panelsArrayToVisibilityMap(o.settings.panels);
-    panels.eval.wrap && wrapEvalOnce();
-    panels.setTimeout.wrap && wrapSetTimeoutOnce();
-    panels.clearTimeout.wrap && wrapClearTimeoutOnce();
-    panels.setInterval.wrap && wrapSetIntervalOnce();
-    panels.clearInterval.wrap && wrapClearIntervalOnce();
-    panels.requestAnimationFrame && wrapRequestAnimationFrameOnce();
-    panels.cancelAnimationFrame && wrapCancelAnimationFrameOnce();
-    panels.requestIdleCallback && wrapRequestIdleCallbackOnce();
-    panels.cancelIdleCallback && wrapCancelIdleCallbackOnce();
-    wrapper.setTraceForDebug(o.settings.traceForDebug);
+    tick.stop();
+    eachSecond.stop();
   } else if (o.msg === 'reset-wrapper-history') {
     wrapper.cleanHistory();
-    tick.trigger();
+    !tick.isPending && tick.trigger();
   } else if (o.msg === 'clear-timer-handler') {
     if (o.type === ETimerType.TIMEOUT) {
       window.clearTimeout(o.handler);
@@ -113,6 +67,11 @@ windowListen((o) => {
     }
   } else if (o.msg === 'media-command') {
     doMediaCommand(o.mediaId, o.cmd, o.property);
+  } else if (o.msg === 'telemetry-acknowledged') {
+    // adaptive update-frequency
+    const ackTrafficDuration = Date.now() - o.timeSent;
+    const newDelay = (o.trafficDuration + ackTrafficDuration) * 3;
+    tick.delay = Math.max(TELEMETRY_FREQUENCY_30PS, newDelay);
   }
 });
 
