@@ -1,4 +1,4 @@
 export const IS_DEV = __development__;
 export const APPLICATION_VERSION = __app_version__;
-export const APPLICATION_NAME = `${__app_name__}@${__app_version__}`;
+export const APPLICATION_NAME = __app_name__;
 export const APPLICATION_HOME_PAGE = __home_page__;
