@@ -1,16 +1,14 @@
 <script lang="ts">
-  import { Fps } from '@/api/time.ts';
+  import { Fps } from '../../api/time.ts';
 
   const FRAMES = '⣷⣯⣟⡿⢿⣻⣽⣾';
-  let index = 0;
-  let frame = FRAMES[index];
-
-  let fpsValue = 0;
+  let index = $state.raw(0);
+  let fpsValue = $state.raw(0);
+  let frame = $derived.by(() => FRAMES[index]);
   const fps = new Fps((value) => (fpsValue = value)).start();
 
   export function tick() {
     index = ++index % FRAMES.length;
-    frame = FRAMES[index];
     fps.tick();
   }
 </script>
@@ -25,5 +23,6 @@
     display: flex;
     align-items: center;
     gap: 0.125rem;
+    line-height: 1;
   }
 </style>
