@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { APPLICATION_HOME_PAGE, APPLICATION_VERSION } from '@/api/env.ts';
+  import { APPLICATION_HOME_PAGE, APPLICATION_VERSION } from '../../api/env.ts';
 </script>
 
 <a target="_blank" href={APPLICATION_HOME_PAGE} title={APPLICATION_HOME_PAGE}
